Use the new xinfo API.
Wed Dec 29 11:37:55 2004  Stefan Westerfeld  <stefan@space.twc.de>

	* gslloader-pat.cc: Use the new xinfo API.

	* bseieee754.h: Fix compiler warnings in bse_ftoi/bse_dtoi on AMD64.

svn path=/trunk/; revision=3026

+++ bse/bseieee754.h
@@ -242,12 +242,12 @@ bse_dtoi (register double f)
 static inline int G_GNUC_CONST
 bse_ftoi (register float v)
 {
-  return v < -0.0 ? v - 0.5 : v + 0.5;
+  return (int) (v < -0.0 ? v - 0.5 : v + 0.5);
 }
 static inline int G_GNUC_CONST
 bse_dtoi (register double v)
 {
-  return v < -0.0 ? v - 0.5 : v + 0.5;
+  return (int) (v < -0.0 ? v - 0.5 : v + 0.5);
 }
 #endif
 static inline guint64
